Get rating by AJAX  call
Adding a constructor to set default state.

Adding the componentDidMount life cycle method to fetch rating data from our local api server that we started in our earlier step.

We make a call and add the data to our state.

We check if apiData is present in the state to show rating. If yes show rating or else show spinner.

I added console.log for debugging.